reviewed and adjusted FreezeEnemies
Freeze enemies was set up correctly but referenced the core movement script.

so reverenced it to the movement script used in test scene for testing, but only commented out the EnemyMovement script for easy transition back to incorporating it.

added Freeze methods to Waypoint class, added If statements to update for proper reference.